Attempt all auth types when reading EK cert
When reading NVRAM values the `NvReadAccessChecks()` func in "Trusted
Platform Module Library Part 4: Supporting Routines" supports owner,
platform, and "self" auth. This change makes it to where go-attestation
attempts all methods with the empty string password.

The main impetus here is to support ChromeOS's vtpm implementation[1],
which doesn't have a concept of an "owner" or "platform" password, and
thus expects to perform the "self" auth for the NVRAM cell containing
the premade EK cert.

[1]: https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform2/vtpm/commands/nv_read_command.cc;l=64-68;drc=1efd0c8f36050d56b8550354a4c7af925e44118a